All ID's have been assigned
The definitions have been completed to the barest minimum necessary for the game. 

Still need to go through the nuclear definitions to make sure the language is understandable for the general public. Then go through the rest of the definitions, Cost definitions could be found under businessdictionary.com. Link any words to Wikipedia or other sites if necessary.